update to Substrate commit 6d82c02611c51a5c54d60e02c3c104b10aa3ae71

+++ runtime/tests/cli_integration_tests_mining_hardware.rs
@@ -80,7 +80,7 @@ mod tests {
 
     parameter_types! {
         pub const BlockHashCount: u64 = 250;
-        pub const SS58Prefix: u8 = 33;
+        pub const SS58Prefix: u16 = 33;
     }
     impl frame_system::Config for Test {
         type AccountData = pallet_balances::AccountData<u64>;
